refactor(CheckOutPriceSection): 컴포넌트 text관련 ui 로직 hook 분리
useCheckOutPriceText hook 생성
